Refactoring, brrrr

+++ src/Joystick.hpp
@@ -2,6 +2,7 @@
 #define __JOYSTICK_HPP__
 
 #include "Controller.hpp"
+#include <stdlib.h>
 
 class joystick : public controller
 {
@@ -14,9 +15,9 @@ public:
   bool right() const;
   bool button() const;
 
-  static constexpr size_t DEF_SENSITIVITY = 900;
+  static constexpr int DEF_SENSITIVITY = 900;
   // this value may differ among different devices
-  static constexpr size_t MAX_ANALOG_VALUE = 1023;
+  static constexpr int MAX_ANALOG_VALUE = 1023;
 
 private:
   int _x, _y, _button, _sensitivity;
